+++ src/storage/ldb/leveldb/db/log_writer.cc
@@ -81,7 +81,7 @@ Status Writer::AddRecord(const Slice& slice) {
 
 Status Writer::EmitPhysicalRecord(RecordType t, const char* ptr, size_t n) {
   assert(n <= 0xffff);  // Must fit in two bytes
-  assert(block_offset_ + kHeaderSize + n <= kBlockSize);
+  assert(block_offset_ + kHeaderSize + n <= (size_t)kBlockSize);
 
   // Format the header
   char buf[kHeaderSize];
